refactor: remove obsolete game store
