+++ src/boardStore.js
@@ -1,4 +1,23 @@
-import React, { createContext, useReducer, useContext, useMemo } from 'react';
+import React, {
+  createContext,
+  useReducer,
+  useContext,
+  useMemo,
+  useState,
+  useEffect,
+  useCallback,
+} from 'react';
+
+const FEN_KEY = 'game_fen';
+const HISTORY_KEY = 'game_history';
+
+function safeJSONParse(str, fallback) {
+  try {
+    return JSON.parse(str);
+  } catch (_) {
+    return fallback;
+  }
+}
 
 function initialBoard() {
   return {
@@ -30,20 +49,63 @@ function reducer(state, action) {
 const BoardContext = createContext(null);
 
 export function BoardProvider({ children }) {
-  const [state, dispatch] = useReducer(reducer, {
+  const [boardState, dispatch] = useReducer(reducer, {
     board: initialBoard(),
     orientation: 'white'
   });
 
+  const [fen, setFen] = useState(() => {
+    if (typeof localStorage !== 'undefined') {
+      return localStorage.getItem(FEN_KEY) || '';
+    }
+    return '';
+  });
+
+  const [history, setHistory] = useState(() => {
+    if (typeof localStorage !== 'undefined') {
+      return safeJSONParse(localStorage.getItem(HISTORY_KEY), []);
+    }
+    return [];
+  });
+
+  useEffect(() => {
+    if (typeof localStorage !== 'undefined') {
+      localStorage.setItem(FEN_KEY, fen);
+    }
+  }, [fen]);
+
+  useEffect(() => {
+    if (typeof localStorage !== 'undefined') {
+      localStorage.setItem(HISTORY_KEY, JSON.stringify(history));
+    }
+  }, [history]);
+
+  const addMove = useCallback(move => {
+    setHistory(h => [...h, move]);
+  }, []);
+
+  const exportPGN = useCallback(() => history.join(' '), [history]);
+
+  const importFEN = useCallback(newFen => {
+    setFen(newFen);
+    setHistory([]);
+  }, []);
+
   const actions = useMemo(() => ({
     playerMove: (from, to) => dispatch({ type: 'PLAYER_MOVE', from, to }),
     aiMove: (from, to) => dispatch({ type: 'AI_MOVE', from, to }),
-    flipOrientation: () => dispatch({ type: 'FLIP_ORIENTATION' })
-  }), [dispatch]);
+    flipOrientation: () => dispatch({ type: 'FLIP_ORIENTATION' }),
+    setFen,
+    addMove,
+    exportPGN,
+    importFEN
+  }), [dispatch, setFen, addMove, exportPGN, importFEN]);
+
+  const state = useMemo(() => ({ ...boardState, fen, history }), [boardState, fen, history]);
 
   const value = useMemo(() => ({ state, actions }), [state, actions]);
 
-  return <BoardContext.Provider value={value}>{children}</BoardContext.Provider>;
+  return React.createElement(BoardContext.Provider, { value }, children);
 }
 
 export function useBoardState() {
